+++ trackeval/datasets/burst_helpers/burst_base.py
@@ -10,7 +10,7 @@ from scipy.optimize import linear_sum_assignment
 
 from trackeval import _timing, utils
 from trackeval.datasets._base_dataset import _BaseDataset
-from trackeval.utils import TrackEvalException
+from trackeval.exception import TrackEvalException
 
 
 class BURSTBase(_BaseDataset):